+++ fsleyes/actions/runscript.py
@@ -207,7 +207,7 @@ def fsleyesScriptEnvironment(frame, overlayList, displayCtx):
         from . import                 loadoverlay
         import fsleyes.autodisplay as autodisplay
 
-        def onLoad(overlays):
+        def onLoad(paths, overlays):
 
             if len(overlays) == 0:
                 return